+++ Biritako-Boom/Assets/Animation/Enemies/Alien.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 95490f81865a80f4dbfaef68c2981828
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
