+++ Biritako-Boom/Assets/Animation/Enemies/Alien/Alien.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 146517fd47f37f74881e33db2ead82be
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
