+++ Biritako-Boom/Assets/Animation/Enemies/Alien/Alien_Idle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 008d86702cf38df4da5159ee4c23dc3b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
