+++ Biritako-Boom/Assets/Animation/Enemies/Alien/Alien_Move.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 84d425531944fb349862c93310a7e268
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
