+++ Biritako-Boom/Assets/Animation/Enemies/MotherShip.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 77e1c94a5a6464d4f9b960d1662b5da1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
