+++ Biritako-Boom/Assets/Animation/Enemies/MotherShip/MotherShip.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b42cfa094382b274bbbef4a0326ab68d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
