+++ Biritako-Boom/Assets/Animation/Enemies/MotherShip/MotherShip_Move.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5bdf717822c489b41afc5a7d21697649
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
