+++ Biritako-Boom/Assets/Prefabs/Enemy_Alien.prefab
@@ -102,6 +102,7 @@ GameObject:
   - component: {fileID: 9189366533288368162}
   - component: {fileID: 4579485241098840379}
   - component: {fileID: 1123496913849891493}
+  - component: {fileID: 6275503577709545430}
   m_Layer: 0
   m_Name: Enemy_Alien
   m_TagString: Untagged
@@ -169,7 +170,7 @@ SpriteRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 2
-  m_Sprite: {fileID: -821724291, guid: ee3d78f1315fa294085af1facfe0c64c, type: 3}
+  m_Sprite: {fileID: -1608743054978694511, guid: fc4bb2110acda554a93564273c799e14, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
@@ -245,7 +246,7 @@ BoxCollider2D:
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
     pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 0.52, y: 0.53}
+    oldSize: {x: 0.47, y: 0.47}
     newSize: {x: 1, y: 1}
     adaptiveTilingThreshold: 0.5
     drawMode: 0
@@ -291,3 +292,25 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 88fc035f9eb34454c8e05bb1f2e67361, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!95 &6275503577709545430
+Animator:
+  serializedVersion: 7
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4894678526528209877}
+  m_Enabled: 1
+  m_Avatar: {fileID: 0}
+  m_Controller: {fileID: 9100000, guid: 146517fd47f37f74881e33db2ead82be, type: 2}
+  m_CullingMode: 0
+  m_UpdateMode: 0
+  m_ApplyRootMotion: 0
+  m_LinearVelocityBlending: 0
+  m_StabilizeFeet: 0
+  m_AnimatePhysics: 0
+  m_WarningMessage: 
+  m_HasTransformHierarchy: 1
+  m_AllowConstantClipSamplingOptimization: 1
+  m_KeepAnimatorStateOnDisable: 0
+  m_WriteDefaultValuesOnDisable: 0
