+++ Biritako-Boom/Assets/Prefabs/Enemy_MotherShip.prefab
@@ -15,6 +15,7 @@ GameObject:
   - component: {fileID: -6851847958364582349}
   - component: {fileID: 1176204740625652298}
   - component: {fileID: 4558666829951707394}
+  - component: {fileID: 8341396105127106121}
   m_Layer: 0
   m_Name: Enemy_MotherShip
   m_TagString: Untagged
@@ -82,7 +83,7 @@ SpriteRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 4
-  m_Sprite: {fileID: -3298995681524961215, guid: 44be2945d557668478c67f4c16678283, type: 3}
+  m_Sprite: {fileID: -7897507592471590671, guid: f7d01ed46a832a848b0d62fcfa1d3404, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
@@ -196,7 +197,7 @@ BoxCollider2D:
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
     pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 1.9, y: 1.89}
+    oldSize: {x: 1.86, y: 1.82}
     newSize: {x: 1, y: 1}
     adaptiveTilingThreshold: 0.5
     drawMode: 0
@@ -204,6 +205,28 @@ BoxCollider2D:
   m_AutoTiling: 0
   m_Size: {x: 1.9, y: 1.89}
   m_EdgeRadius: 0
+--- !u!95 &8341396105127106121
+Animator:
+  serializedVersion: 7
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1507075229319319441}
+  m_Enabled: 1
+  m_Avatar: {fileID: 0}
+  m_Controller: {fileID: 9100000, guid: b42cfa094382b274bbbef4a0326ab68d, type: 2}
+  m_CullingMode: 0
+  m_UpdateMode: 0
+  m_ApplyRootMotion: 0
+  m_LinearVelocityBlending: 0
+  m_StabilizeFeet: 0
+  m_AnimatePhysics: 0
+  m_WarningMessage: 
+  m_HasTransformHierarchy: 1
+  m_AllowConstantClipSamplingOptimization: 1
+  m_KeepAnimatorStateOnDisable: 0
+  m_WriteDefaultValuesOnDisable: 0
 --- !u!1 &5430154999915309962
 GameObject:
   m_ObjectHideFlags: 0
