+++ Biritako-Boom/Assets/Scripts/Common/CameraManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cfee416853596c748a7988f334680a1a
