+++ Biritako-Boom/Assets/Scripts/Common/ShakeEffect/IShakeParameters.cs.meta
