+++ Biritako-Boom/Assets/Scripts/Common/ShakeEffect/ShakeInstance.cs
