+++ Biritako-Boom/Assets/Scripts/Common/ShakeEffect/ShakeParameters.cs
