+++ Biritako-Boom/Assets/Scripts/Common/ShakeEffect/ShakePreset.cs
