+++ Biritako-Boom/Assets/Scripts/Common/ShakeEffect/ShakeResult.cs
