+++ Biritako-Boom/Assets/Scripts/Common/ShakeEffect/ShakeState.cs
