+++ Biritako-Boom/Assets/Scripts/Common/ShakeEffect/ShakeType.cs
