--- Biritako-Boom/Assets/Scripts/InGame/Model/CameraModel.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2412ba7b6b21cab479ccf2426eb13c4e
