--- Biritako-Boom/Assets/Scripts/InGame/Presenter/CameraPresenter.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d9b10af9c7056174ca8e25df4cf9826d
