--- Biritako-Boom/Assets/Scripts/InGame/View/CameraView.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3a61e93418e909a4f8416e5aaa4ab89c
