+++ Biritako-Boom/Assets/Scripts/Shader.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6f1b39fe1b4edfa48b5bf0f78c4a47ed
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
